codex-tools: extract local host tool specs
